+++ frontend/package.json
@@ -16,6 +16,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^5.0.1",
+    "react-router-dom": "^6.22.1",
     "react-scripts": "5.0.1",
     "sass": "^1.70.0",
     "typescript": "^4.9.5",
