+++ frontend/src/App.module.scss
@@ -2,13 +2,5 @@
   display: flex;
   flex-direction: column;
   justify-content: space-between;
-  height: 100vh;
-}
-
-.cards__cont {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  margin: 50px 4%;
-  gap: 20px;
+  min-height: 100vh;
 }
