+++ frontend/src/pages/IndexPage/IndexPage.module.scss
@@ -1,0 +1,7 @@
+.cards__cont {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+  margin: 50px 4%;
+  gap: 20px;
+}
